Cavity Map -JS
You are given a square map of size . Each cell of the map has a value denoting its depth. We will call a cell of the map a cavity if and only if this cell is not on the border of the map and each cell adjacent to it has strictly smaller depth. Two cells are adjacent if they have a common side (edge).

You need to find all the cavities on the map and depict them with the uppercase character X.

Input Format

The first line contains an integer, , denoting the size of the map. Each of the following lines contains positive digits without spaces. Each digit (1-9) denotes the depth of the appropriate area.

Constraints

Output Format

Output lines, denoting the resulting map. Each cavity should be replaced with character X.